chore(refactor): split Home.tsx into a Home/ directory of siblings
The 1923-line Home.tsx had grown 11 inner sub-components and 7 helper
functions tacked on after the main Home component. None of them reached
into Home's outer scope beyond their props, so they're cleanly siblings.

Layout:
  web/src/components/Home/
    index.tsx              the Home component (still the named export)
    types.ts               ViewMode / StateFilter / ArtefactSource / VAULT
    utils.tsx              homeRelative / renderPipCounts / stateColor /
                           metaForSession / formatRelative / pluralize /
                           formatBytes / spaceLabelFor + AGENT_* maps
    SessionTile.tsx
    SessionRow.tsx
    ArtefactTable.tsx
    ShowMore.tsx
    AddMemoryForm.tsx
    ProjectTileGrid.tsx
    ProjectTile.tsx        (rendered by ProjectTileGrid)
    AttachFolderForm.tsx
    MemoryCard.tsx
    VaultInfo.tsx
    VaultContents.tsx      (rendered by VaultInfo)
    Home.css               (moved from web/src/components/Home.css)

Home.tsx and Home.css became Home/index.tsx and Home/Home.css; the
external import path (App.tsx → "./components/Home") still resolves
because TypeScript walks to index.tsx.

Things kept inline in index.tsx because they're tightly coupled to
Home's own state / single-caller surface:
  - useStickyView (single caller)
  - ARTEFACT_SOURCE_ORDER / ARTEFACT_SOURCE_LABELS / FILTER_ORDER /
    FILTER_LABELS / LIVE_STATES / EMPTY_COUNTS / MEMORIES_PREVIEW /
    ARTEFACTS_PREVIEW (Home-only constants)

LOC: index.tsx 1923 → 1114 lines (-809, -42%). Sibling files total
~870 lines plus 11 lines of shared types — ~57 lines of per-module
ceremony (imports, type re-exports). Net +59 lines across the directory
in exchange for navigable file boundaries.

No behavioural change.
