+++ web/src/components/Home/index.tsx
@@ -1,22 +1,29 @@
 import { useEffect, useMemo, useRef, useState } from "react";
 import { LayoutGroup, motion } from "framer-motion";
-import { ArrowUpRight, Brain, Folder, FolderPlus, Globe, RefreshCw, Shield } from "lucide-react";
-import type { Session, SessionState, SessionAgent } from "../data/sessions-api";
-import type { Memory } from "../data/memories-api";
-import { createMemory } from "../data/memories-api";
-import type { Space } from "../../../shared/types";
-import { useSessions } from "../hooks/useSessions";
-import { useMemories } from "../hooks/useMemories";
-import { useSpaceSources } from "../hooks/useSpaceSources";
-import type { SpaceSource } from "../data/spaces-api";
-import { addSpaceSource, removeSpaceSource } from "../data/spaces-api";
-import { parseTimestamp } from "../utils/parseTimestamp";
-import { Desktop } from "./Desktop";
-import { InspectorPanel, type ActivePanel } from "./InspectorPanel";
-import { SessionInspector } from "./SessionInspector";
-import { ArtefactInspector } from "./ArtefactInspector";
-import { ConfirmModal } from "./ConfirmModal";
-import { SpaceContextMenu } from "./SpaceContextMenu";
+import { ArrowUpRight, Folder, FolderPlus, Shield } from "lucide-react";
+import type { SessionState } from "../../data/sessions-api";
+import type { Space } from "../../../../shared/types";
+import { useSessions } from "../../hooks/useSessions";
+import { useMemories } from "../../hooks/useMemories";
+import { useSpaceSources } from "../../hooks/useSpaceSources";
+import { parseTimestamp } from "../../utils/parseTimestamp";
+import { Desktop } from "../Desktop";
+import { InspectorPanel, type ActivePanel } from "../InspectorPanel";
+import { SessionInspector } from "../SessionInspector";
+import { ArtefactInspector } from "../ArtefactInspector";
+import { ConfirmModal } from "../ConfirmModal";
+import { SpaceContextMenu } from "../SpaceContextMenu";
+import { SessionTile } from "./SessionTile";
+import { SessionRow } from "./SessionRow";
+import { ArtefactTable } from "./ArtefactTable";
+import { ShowMore } from "./ShowMore";
+import { AddMemoryForm } from "./AddMemoryForm";
+import { ProjectTileGrid } from "./ProjectTileGrid";
+import { AttachFolderForm } from "./AttachFolderForm";
+import { MemoryCard } from "./MemoryCard";
+import { VaultInfo } from "./VaultInfo";
+import { homeRelative, renderPipCounts, stateColor } from "./utils";
+import { VAULT, type ArtefactSource, type StateFilter, type ViewMode } from "./types";
 import "./Home.css";
 
 interface Props {
@@ -33,10 +40,6 @@ interface Props {
   /** Used by the breadcrumb-pill context menu (rename). */
   onSpaceUpdate?: (id: string, fields: { displayName?: string; color?: string }) => void;
 }
-
-type ViewMode = "icons" | "table";
-type StateFilter = SessionState | "live" | "all";
-type ArtefactSource = "all" | "manual" | "ai_generated" | "discovered";
 
 const ARTEFACT_SOURCE_ORDER: ArtefactSource[] = ["all", "manual", "ai_generated", "discovered"];
 const ARTEFACT_SOURCE_LABELS: Record<ArtefactSource, string> = {
@@ -92,10 +95,6 @@ const EMPTY_COUNTS = { total: 0, active: 0, waiting: 0, disconnected: 0, done: 0
 // <5 memories) stay fully visible.
 const MEMORIES_PREVIEW = 5;
 
-// Sentinel for the Vault tile (artefacts with no source_id —
-// natively-created via create_artifact, not from a linked folder).
-const VAULT = "__vault__";
-
 const FILTER_LABELS: Record<StateFilter, string> = {
   live: "live",
   active: "active",
@@ -104,50 +103,6 @@ const FILTER_LABELS: Record<StateFilter, string> = {
   done: "done",
   all: "all",
 };
-
-const AGENT_LETTERS: Record<SessionAgent, string> = {
-  "claude-code": "CC",
-  opencode: "OC",
-  codex: "CX",
-};
-
-const AGENT_CLASS: Record<SessionAgent, string> = {
-  "claude-code": "agent-cc",
-  opencode: "agent-oc",
-  codex: "agent-codex",
-};
-
-const AGENT_PIP_CLASS: Record<SessionAgent, string> = {
-  "claude-code": "cc",
-  opencode: "oc",
-  codex: "codex",
-};
-
-// Collapse `/Users/<name>/...` and `/home/<name>/...` to `~/...` so
-// orphan-cwd labels read at a glance. Falls through unchanged for
-// Windows paths and anything outside the user home tree.
-function homeRelative(p: string): string {
-  const m = p.match(/^\/(?:Users|home)\/[^/]+/);
-  return m ? "~" + p.slice(m[0].length) : p;
-}
-
-// Compact pip + numeral, one per non-zero state. Same glow primitive
-// the Active-projects tile signals use, just without the trailing
-// state word — fits in a breadcrumb pill while keeping counts exact
-// and the visual language consistent with the tiles.
-function renderPipCounts(counts: { active?: number; waiting?: number; disconnected?: number }) {
-  const a = counts.active ?? 0;
-  const w = counts.waiting ?? 0;
-  const d = counts.disconnected ?? 0;
-  if (a + w + d === 0) return null;
-  return (
-    <>
-      {a > 0 && <span className="pip-count"><span className="pip pip-green" />{a}</span>}
-      {w > 0 && <span className="pip-count"><span className="pip pip-amber" />{w}</span>}
-      {d > 0 && <span className="pip-count"><span className="pip pip-red" />{d}</span>}
-    </>
-  );
-}
 
 export function Home({ activeSpace, spaces, desktopProps, isHero, onSpaceChange, onPromoteFolderToSpace, onSpaceDelete, onSpaceUpdate }: Props) {
   const { sessions, error, loading } = useSessions();
@@ -1157,767 +1112,3 @@ export function Home({ activeSpace, spaces, desktopProps, isHero, onSpaceChange,
   );
 }
 
-function stateColor(state: SessionState): "green" | "amber" | "red" | "dim" {
-  switch (state) {
-    case "active": return "green";
-    case "waiting": return "amber";
-    case "disconnected": return "red";
-    case "done": return "dim";
-  }
-}
-
-function spaceLabelFor(spaceId: string | null, spaces: Space[]): string | null {
-  if (!spaceId) return null;
-  return spaces.find((s) => s.id === spaceId)?.displayName ?? spaceId;
-}
-
-function metaForSession(session: Session): string {
-  const rel = formatRelative(session.lastEventAt) ?? "—";
-  if (session.state === "waiting") return `${session.agent} · waiting ${rel}`;
-  if (session.state === "disconnected") return `${session.agent} · disconnected ${rel}`;
-  return `${session.agent} · ${rel}`;
-}
-
-interface SessionTileProps {
-  session: Session;
-  spaces: Space[];
-  showSpaceChip: boolean;
-  onOpen?: (id: string) => void;
-}
-
-function SessionTile({ session, spaces, showSpaceChip, onOpen }: SessionTileProps) {
-  const spaceLabel = spaceLabelFor(session.spaceId, spaces);
-  const title = session.title ?? "(no title yet)";
-  return (
-    <div
-      className="home-tile"
-      onClick={() => onOpen?.(session.id)}
-      onKeyDown={onOpen ? (e) => {
-        if (e.key === "Enter" || e.key === " ") { e.preventDefault(); onOpen(session.id); }
-      } : undefined}
-      role={onOpen ? "button" : undefined}
-      tabIndex={onOpen ? 0 : undefined}
-    >
-      <div className={`home-thumb ${AGENT_CLASS[session.agent]}`}>
-        {showSpaceChip && spaceLabel && (
-          <span className="home-space-chip">{spaceLabel}</span>
-        )}
-        <span className="home-agent-mark">{AGENT_LETTERS[session.agent]}</span>
-        <span className={`home-status ${session.state}`} />
-      </div>
-      <div className="home-tile-label" title={title}>{title}</div>
-      <div className="home-tile-meta">{metaForSession(session)}</div>
-    </div>
-  );
-}
-
-interface SessionRowProps {
-  session: Session;
-  spaces: Space[];
-  onOpen?: (id: string) => void;
-}
-
-function SessionRow({ session, spaces, onOpen }: SessionRowProps) {
-  const spaceLabel = spaceLabelFor(session.spaceId, spaces);
-  const rel = formatRelative(session.lastEventAt) ?? "—";
-  const time = session.state === "waiting" ? `waiting ${rel}`
-    : session.state === "disconnected" ? `disconnected ${rel}`
-    : rel;
-  const title = session.title ?? "(no title yet)";
-  // Prefer the most specific label available: source (folder) > space >
-  // cwd basename for orphan sessions. Always tooltip the full cwd so
-  // the user can identify where the session was running.
-  const cwdBasename = session.cwd ? session.cwd.split(/[\\/]/).filter(Boolean).pop() ?? null : null;
-  const projectLabel = session.sourceLabel ?? spaceLabel ?? cwdBasename ?? "—";
-  return (
-    <div
-      className="home-row"
-      onClick={() => onOpen?.(session.id)}
-      onKeyDown={onOpen ? (e) => {
-        if (e.key === "Enter" || e.key === " ") { e.preventDefault(); onOpen(session.id); }
-      } : undefined}
-      role={onOpen ? "button" : undefined}
-      tabIndex={onOpen ? 0 : undefined}
-    >
-      <span className={`home-row-status ${session.state}`} />
-      <span className="home-row-space" title={session.cwd ?? undefined}>{projectLabel}</span>
-      <span className="home-row-title" title={title}>{title}</span>
-      <span className={`home-row-agent ${AGENT_PIP_CLASS[session.agent]}`}>
-        <span className="home-agent-pip" />
-        {session.agent}
-      </span>
-      <span className="home-row-time">{time}</span>
-    </div>
-  );
-}
-
-interface ArtefactTableProps {
-  artifacts: Parameters<typeof Desktop>[0]["artifacts"];
-  spaces: Space[];
-  onArtifactClick: Parameters<typeof Desktop>[0]["onArtifactClick"];
-}
-
-function ArtefactTable({ artifacts, spaces, onArtifactClick }: ArtefactTableProps) {
-  if (artifacts.length === 0) {
-    return <div className="home-empty">No artefacts here yet.</div>;
-  }
-  const sorted = [...artifacts].sort((a, b) => {
-    const ta = parseTimestamp(a.createdAt);
-    const tb = parseTimestamp(b.createdAt);
-    return (Number.isFinite(tb) ? tb : 0) - (Number.isFinite(ta) ? ta : 0);
-  });
-  return (
-    <div className="home-table-wrap">
-      <div className="home-table">
-        {sorted.map((art) => {
-          const space = spaces.find((s) => s.id === art.spaceId);
-          return (
-            <div
-              key={art.id}
-              className="home-artefact-row"
-              role="button"
-              tabIndex={0}
-              onClick={() => onArtifactClick(art)}
-              onKeyDown={(e) => {
-                if (e.key === "Enter" || e.key === " ") {
-                  e.preventDefault();
-                  onArtifactClick(art);
-                }
-              }}
-            >
-              <span className="home-artefact-row-title">{art.label}</span>
-              <span className="home-artefact-row-space">{space?.displayName ?? art.spaceId}</span>
-              <span className="home-artefact-row-kind">{art.artifactKind}</span>
-              <span className="home-artefact-row-time">{formatRelative(art.createdAt) ?? "—"}</span>
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-// Reused under both the Memories and Artefacts sections. "Show more"
-// loads another preview-sized batch (5 memories, 21 artefacts). The
-// optional ⌘K hint only renders where Spotlight actually searches —
-// artifacts today, eventually memories + sessions when #264 ships.
-function ShowMore({
-  onClick, remaining, searchHint = false,
-}: { onClick: () => void; remaining: number; searchHint?: boolean }) {
-  return (
-    <div className="home-show-more">
-      <button type="button" className="home-memories-toggle" onClick={onClick}>
-        Show more
-      </button>
-      <span className="home-show-more-hint">
-        {remaining} more
-        {searchHint && (
-          <>
-            {" · "}<kbd>⌘K</kbd> to search
-          </>
-        )}
-      </span>
-    </div>
-  );
-}
-
-interface AddMemoryFormProps {
-  defaultSpaceId: string | null;
-  spaces: Space[];
-  onSaved: () => void;
-  onCancel: () => void;
-}
-
-function AddMemoryForm({ defaultSpaceId, spaces, onSaved, onCancel }: AddMemoryFormProps) {
-  const [content, setContent] = useState("");
-  const [spaceId, setSpaceId] = useState<string>(defaultSpaceId ?? "");
-  const [tagsInput, setTagsInput] = useState("");
-  const [submitting, setSubmitting] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const realSpaces = useMemo(
-    () => spaces.filter((s) => s.id !== "home" && s.id !== "__all__" && s.id !== "__archived__"),
-    [spaces],
-  );
-
-  async function submit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!content.trim() || submitting) return;
-    setSubmitting(true);
-    setError(null);
-    try {
-      const tags = tagsInput
-        .split(",")
-        .map((t) => t.trim())
-        .filter(Boolean);
-      await createMemory({
-        content: content.trim(),
-        space_id: spaceId || undefined,
-        tags: tags.length ? tags : undefined,
-      });
-      onSaved();
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  return (
-    <form className="home-memories-add" onSubmit={submit}>
-      <textarea
-        className="home-memories-add-text"
-        placeholder="What should I remember?"
-        value={content}
-        onChange={(e) => setContent(e.target.value)}
-        rows={3}
-        autoFocus
-      />
-      <div className="home-memories-add-row">
-        <select
-          className="home-memories-add-select"
-          value={spaceId}
-          onChange={(e) => setSpaceId(e.target.value)}
-        >
-          <option value="">No space (global)</option>
-          {realSpaces.map((s) => (
-            <option key={s.id} value={s.id}>{s.displayName}</option>
-          ))}
-        </select>
-        <input
-          className="home-memories-add-tags"
-          placeholder="tags, comma-separated"
-          value={tagsInput}
-          onChange={(e) => setTagsInput(e.target.value)}
-        />
-        <div className="home-memories-add-actions">
-          <button type="button" className="home-memories-add-cancel" onClick={onCancel}>
-            Cancel
-          </button>
-          <button
-            type="submit"
-            className="home-memories-add-save"
-            disabled={!content.trim() || submitting}
-          >
-            {submitting ? "Saving…" : "Save memory"}
-          </button>
-        </div>
-      </div>
-      {error && <div className="home-memories-add-error">Couldn't save: {error}</div>}
-    </form>
-  );
-}
-
-// Project tile grid — same visual primitive as Home's space cards.
-// Renders one tile per attached folder (plus an "All" meta-tile, a
-// Vault tile for native artefacts, and a "+ Attach" tile). The
-// selected tile is exclusive: clicking another switches scope, clicking
-// the selected tile snaps back to All. Detach lives in a hover ⋯ menu
-// on linked tiles only — Vault can't be detached.
-function ProjectTileGrid({
-  spaceId, spaceDisplayName, sources, folderArtefactCounts, sessionCountsBySource,
-  selectedFolderId, setSelectedFolderId,
-  totalCounts, showAttachForm, setShowAttachForm, onSourcesChanged, onSpaceDelete,
-}: {
-  spaceId: string;
-  spaceDisplayName: string;
-  sources: SpaceSource[];
-  folderArtefactCounts: Record<string, number>;
-  sessionCountsBySource: Record<string, { active: number; waiting: number; disconnected: number }>;
-  selectedFolderId: string | null;
-  setSelectedFolderId: (next: string | null) => void;
-  totalCounts: Record<StateFilter, number>;
-  showAttachForm: boolean;
-  setShowAttachForm: (v: boolean) => void;
-  onSourcesChanged: () => void;
-  onSpaceDelete?: (spaceId: string) => Promise<void> | void;
-}) {
-  // Sort linked tiles by tile count desc — busiest folders first.
-  const sortedSources = useMemo(
-    () => [...sources].sort((a, b) =>
-      (folderArtefactCounts[b.id] ?? 0) - (folderArtefactCounts[a.id] ?? 0)
-    ),
-    [sources, folderArtefactCounts],
-  );
-  const vaultCount = folderArtefactCounts[VAULT] ?? 0;
-
-  function pickTile(id: string | null) {
-    setSelectedFolderId(selectedFolderId === id ? null : id);
-  }
-
-  return (
-    <div className="home-spaces-section home-projects-section">
-      <div className="home-spaces-grid">
-        <button
-          type="button"
-          className={`home-space-card${selectedFolderId === null ? " selected" : ""}`}
-          onClick={() => pickTile(null)}
-          title="All projects in this space"
-        >
-          <div className="home-space-card-name">All</div>
-          <div className="home-space-card-counts">
-            {totalCounts.active > 0 && <span className="signal"><span className="pip pip-green" />{totalCounts.active} active</span>}
-            {totalCounts.waiting > 0 && <span className="signal"><span className="pip pip-amber" />{totalCounts.waiting} waiting</span>}
-            {totalCounts.disconnected > 0 && <span className="signal"><span className="pip pip-red" />{totalCounts.disconnected} disconnected</span>}
-            {totalCounts.done > 0 && <span className="signal"><span className="pip pip-dim" />{totalCounts.done} done</span>}
-            {totalCounts.all === 0 && <span className="signal signal-muted">no sessions yet</span>}
-          </div>
-        </button>
-
-        {vaultCount > 0 && (
-          <button
-            type="button"
-            className={`home-space-card home-project-tile--vault${selectedFolderId === VAULT ? " selected" : ""}`}
-            onClick={() => pickTile(VAULT)}
-            title="Native artefacts created in this space (not from a linked folder)"
-          >
-            <div className="home-space-card-name">
-              <Shield size={12} strokeWidth={2} fill="currentColor" aria-hidden="true" className="home-project-glyph" />
-              <span>{spaceId}</span>
-              <span className="home-project-tag">vault</span>
-            </div>
-            <div className="home-space-card-counts">
-              <span className="signal"><span className="pip pip-dim" />{vaultCount} {vaultCount === 1 ? "artefact" : "artefacts"}</span>
-            </div>
-          </button>
-        )}
-
-        {sortedSources.map((s) => (
-          <ProjectTile
-            key={s.id}
-            source={s}
-            artefactCount={folderArtefactCounts[s.id] ?? 0}
-            sessionCounts={sessionCountsBySource[s.id]}
-            selected={selectedFolderId === s.id}
-            onSelect={() => pickTile(s.id)}
-            onSourcesChanged={onSourcesChanged}
-            isLastSource={sources.length === 1}
-            spaceDisplayName={spaceDisplayName}
-            spaceTotalSessions={totalCounts.all}
-            onSpaceDelete={onSpaceDelete}
-          />
-        ))}
-
-        <button
-          type="button"
-          className="home-space-card home-project-tile--add"
-          onClick={() => setShowAttachForm(true)}
-        >
-          <div className="home-space-card-name">+ Attach folder</div>
-          <div className="home-space-card-counts">
-            <span className="signal signal-muted">link a repo or folder</span>
-          </div>
-        </button>
-      </div>
-
-      {showAttachForm && (
-        <AttachFolderForm
-          spaceId={spaceId}
-          onAttached={() => {
-            setShowAttachForm(false);
-            onSourcesChanged();
-          }}
-          onCancel={() => setShowAttachForm(false)}
-        />
-      )}
-    </div>
-  );
-}
-
-function ProjectTile({
-  source, artefactCount, sessionCounts, selected, onSelect, onSourcesChanged,
-  isLastSource, spaceDisplayName, spaceTotalSessions, onSpaceDelete,
-}: {
-  source: SpaceSource;
-  artefactCount: number;
-  sessionCounts?: { active: number; waiting: number; disconnected: number };
-  selected: boolean;
-  onSelect: () => void;
-  onSourcesChanged: () => void;
-  /** True when this is the only source in the space — removing it collapses the space. */
-  isLastSource: boolean;
-  spaceDisplayName: string;
-  spaceTotalSessions: number;
-  onSpaceDelete?: (spaceId: string) => Promise<void> | void;
-}) {
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [confirmOpen, setConfirmOpen] = useState(false);
-  const [busy, setBusy] = useState(false);
-  // Separator-agnostic so Windows paths (`C:\Users\...`) display correctly too.
-  const basename = source.path.split(/[\\/]/).filter(Boolean).pop() ?? source.path;
-  // Removing the only folder from a real space implicitly demotes the space:
-  // we soft-delete the source then ask App to delete the (now empty) space.
-  // Cascade returns sessions to Elsewhere via sessions.space_id ON DELETE SET NULL.
-  const willCollapseSpace = isLastSource && Boolean(onSpaceDelete);
-
-  // Close menu on outside click — same pattern as space-card menus.
-  useEffect(() => {
-    if (!menuOpen) return;
-    const onDocClick = (e: MouseEvent) => {
-      if (!(e.target as HTMLElement)?.closest(".home-project-tile-menu, .home-project-tile-more")) {
-        setMenuOpen(false);
-      }
-    };
-    document.addEventListener("click", onDocClick);
-    return () => document.removeEventListener("click", onDocClick);
-  }, [menuOpen]);
-
-  async function performRemove() {
-    setBusy(true);
-    try {
-      await removeSpaceSource(source.space_id, source.id);
-      if (willCollapseSpace) {
-        await onSpaceDelete!(source.space_id);
-      }
-      onSourcesChanged();
-      setConfirmOpen(false);
-    } catch (err) {
-      alert(`Couldn't remove: ${err instanceof Error ? err.message : String(err)}`);
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  const sessionPhrase = spaceTotalSessions === 1 ? "1 session" : `${spaceTotalSessions} sessions`;
-
-  return (
-    <>
-      <div className={`home-space-card home-project-tile${selected ? " selected" : ""}`}>
-        <button
-          type="button"
-          className="home-project-tile-body"
-          onClick={onSelect}
-          title={source.path}
-        >
-          <div className="home-space-card-name">{basename}</div>
-          <div className="home-space-card-counts">
-            {sessionCounts && sessionCounts.active > 0 && <span className="signal"><span className="pip pip-green" />{sessionCounts.active} active</span>}
-            {sessionCounts && sessionCounts.waiting > 0 && <span className="signal"><span className="pip pip-amber" />{sessionCounts.waiting} waiting</span>}
-            {sessionCounts && sessionCounts.disconnected > 0 && <span className="signal"><span className="pip pip-red" />{sessionCounts.disconnected} disconnected</span>}
-            <span className="signal"><span className="pip pip-dim" />{artefactCount} {artefactCount === 1 ? "artefact" : "artefacts"}</span>
-          </div>
-        </button>
-        <button
-          type="button"
-          className={`home-project-tile-more${menuOpen ? " open" : ""}`}
-          onClick={(e) => { e.stopPropagation(); setMenuOpen((v) => !v); }}
-          aria-label="Folder actions"
-          aria-haspopup="menu"
-          aria-expanded={menuOpen}
-        >
-          ⋯
-        </button>
-        {menuOpen && (
-          <div className="home-project-tile-menu" role="menu">
-            <div className="home-project-tile-menu-path">{source.path}</div>
-            <div className="home-project-tile-menu-divider" />
-            <button
-              type="button"
-              className="home-project-tile-menu-item danger"
-              onClick={() => { setMenuOpen(false); setConfirmOpen(true); }}
-            >
-              Remove folder…
-            </button>
-          </div>
-        )}
-      </div>
-      <ConfirmModal
-        open={confirmOpen}
-        title={willCollapseSpace ? `Remove "${basename}" and delete ${spaceDisplayName}?` : `Remove "${basename}"?`}
-        body={
-          <>
-            <div style={{ fontFamily: "var(--font-mono)", fontSize: 12, color: "var(--text-dim)", marginBottom: 8 }}>
-              {source.path}
-            </div>
-            {willCollapseSpace ? (
-              <>
-                This is the only folder in <strong>{spaceDisplayName}</strong>.
-                Removing it will delete the space, hide its artefacts, and send {sessionPhrase} back to Elsewhere.
-              </>
-            ) : (
-              <>
-                Its artefacts will be hidden. Sessions stay in <strong>{spaceDisplayName}</strong>.
-                Reattach the same path to restore them.
-              </>
-            )}
-          </>
-        }
-        confirmLabel={busy ? (willCollapseSpace ? "Deleting…" : "Removing…") : (willCollapseSpace ? "Remove and delete" : "Remove")}
-        destructive
-        onConfirm={performRemove}
-        onCancel={() => !busy && setConfirmOpen(false)}
-      />
-    </>
-  );
-}
-
-function AttachFolderForm({
-  spaceId, onAttached, onCancel,
-}: {
-  spaceId: string;
-  onAttached: () => void;
-  onCancel: () => void;
-}) {
-  const [path, setPath] = useState("");
-  const [submitting, setSubmitting] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  async function submit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!path.trim() || submitting) return;
-    setSubmitting(true);
-    setError(null);
-    try {
-      await addSpaceSource(spaceId, path.trim());
-      onAttached();
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  return (
-    <form className="home-memories-add" onSubmit={submit}>
-      <input
-        className="home-memories-add-text"
-        style={{ minHeight: 0 }}
-        placeholder="/absolute/path/to/folder (or ~/path)"
-        value={path}
-        onChange={(e) => setPath(e.target.value)}
-        autoFocus
-      />
-      <div className="home-memories-add-row">
-        <span className="home-memories-add-error" style={{ flex: 1, color: "var(--text-dim)" }}>
-          The folder will be scanned in the background.
-        </span>
-        <div className="home-memories-add-actions">
-          <button type="button" className="home-memories-add-cancel" onClick={onCancel}>
-            Cancel
-          </button>
-          <button
-            type="submit"
-            className="home-memories-add-save"
-            disabled={!path.trim() || submitting}
-          >
-            {submitting ? "Attaching…" : "Attach"}
-          </button>
-        </div>
-      </div>
-      {error && <div className="home-memories-add-error">Couldn't attach: {error}</div>}
-    </form>
-  );
-}
-
-interface MemoryCardProps {
-  memory: Memory;
-  spaces: Space[];
-  showSpaceChip: boolean;
-  onOpenSession: (id: string) => void;
-}
-
-function MemoryCard({ memory, spaces, showSpaceChip, onOpenSession }: MemoryCardProps) {
-  const spaceLabel = spaceLabelFor(memory.space_id, spaces);
-  const rel = formatRelative(memory.created_at) ?? "—";
-  return (
-    <div className="home-memory">
-      <div className="home-memory-text">{memory.content}</div>
-      <div className="home-memory-meta">
-        {showSpaceChip && spaceLabel && <span className="home-memory-space">{spaceLabel}</span>}
-        {memory.tags.length > 0 && (
-          <span className="home-memory-tags">
-            {memory.tags.map((t) => (
-              <span key={t} className="home-memory-tag">{t}</span>
-            ))}
-          </span>
-        )}
-        {memory.source_session_id && (
-          <button
-            type="button"
-            className="home-memory-source"
-            onClick={() => onOpenSession(memory.source_session_id!)}
-            title="Open the session that wrote this memory"
-          >
-            from session
-          </button>
-        )}
-        <span className="home-memory-time">{rel}</span>
-      </div>
-    </div>
-  );
-}
-
-function formatRelative(iso: string): string | null {
-  const t = parseTimestamp(iso);
-  if (!Number.isFinite(t)) return null;
-  const ms = Date.now() - t;
-  if (ms < 0) return "just now";
-  const s = Math.floor(ms / 1000);
-  if (s < 60) return `${s}s ago`;
-  const m = Math.floor(s / 60);
-  if (m < 60) return `${m}m`;
-  const h = Math.floor(m / 60);
-  if (h < 24) return `${h}h`;
-  const d = Math.floor(h / 24);
-  if (d < 30) return `${d}d`;
-  return new Date(t).toLocaleDateString();
-}
-
-function VaultInfo() {
-  return (
-    <div className="home-vault-page">
-      <section className="home-vault-hero">
-        <div className="home-vault-hero-glow" aria-hidden="true" />
-        <div className="home-vault-hero-content">
-          <div className="home-vault-hero-eyebrow">
-            <Shield size={14} strokeWidth={2} fill="currentColor" aria-hidden="true" />
-            <span>Oyster Pro</span>
-            <span className="home-vault-hero-pill">Coming soon</span>
-          </div>
-          <h2 className="home-vault-hero-title">Your Oyster, everywhere.</h2>
-          <p className="home-vault-hero-sub">
-            Your spaces, artifacts, and your AI's memory — continue anywhere,
-            in any agent. Backed up, encrypted, picked up wherever you left off.
-          </p>
-          <div className="home-vault-hero-chips">
-            <span className="home-vault-chip">
-              <RefreshCw size={12} strokeWidth={2} aria-hidden="true" />
-              Sync
-            </span>
-            <span className="home-vault-chip">
-              <Brain size={12} strokeWidth={2} aria-hidden="true" />
-              Memory
-            </span>
-            <span className="home-vault-chip">
-              <Globe size={12} strokeWidth={2} aria-hidden="true" />
-              Publish
-            </span>
-          </div>
-          <div className="home-vault-hero-cta">
-            <a
-              className="home-vault-hero-button"
-              href="https://oyster.to/pricing#waitlist"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Join the waitlist
-            </a>
-            <a
-              className="home-vault-hero-button home-vault-hero-button--secondary"
-              href="https://oyster.to/pricing"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Read more
-            </a>
-          </div>
-        </div>
-        <div className="home-vault-hero-art" aria-hidden="true">
-          <div className="home-vault-hero-art-ring" />
-          <div className="home-vault-hero-art-ring home-vault-hero-art-ring--mid" />
-          <div className="home-vault-hero-art-ring home-vault-hero-art-ring--inner" />
-          <Shield size={56} strokeWidth={1.5} className="home-vault-hero-art-shield" />
-        </div>
-      </section>
-
-      <VaultContents />
-
-    </div>
-  );
-}
-
-interface VaultInventoryEntry {
-  name: string;
-  label: string;
-  description: string;
-  count: number;
-  unit: string;
-  size: number;
-  exists: boolean;
-  meta?: string;
-}
-
-interface VaultInventory {
-  root: string;
-  totalSize: number;
-  entries: VaultInventoryEntry[];
-}
-
-function pluralize(n: number, unit: string): string {
-  if (n === 1) return unit;
-  if (unit.endsWith("y")) return unit.slice(0, -1) + "ies";
-  return unit + "s";
-}
-
-function VaultContents() {
-  const [inv, setInv] = useState<VaultInventory | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  useEffect(() => {
-    let cancelled = false;
-    fetch("/api/vault/inventory")
-      .then((r) => r.ok ? r.json() : Promise.reject(new Error(`HTTP ${r.status}`)))
-      .then((data) => { if (!cancelled) setInv(data); })
-      .catch((err) => { if (!cancelled) setError(err instanceof Error ? err.message : String(err)); });
-    return () => { cancelled = true; };
-  }, []);
-
-  if (error) {
-    return (
-      <section className="home-vault-contents">
-        <div className="home-vault-contents-head">
-          <span className="home-vault-contents-label">Your Oyster</span>
-          <span className="home-vault-section-rule" />
-        </div>
-        <p className="home-vault-empty">Couldn't load: {error}</p>
-      </section>
-    );
-  }
-  if (!inv) {
-    return (
-      <section className="home-vault-contents">
-        <div className="home-vault-contents-head">
-          <span className="home-vault-contents-label">Your Oyster</span>
-          <span className="home-vault-section-rule" />
-        </div>
-        <p className="home-vault-empty">Reading your vault…</p>
-      </section>
-    );
-  }
-  return (
-    <section className="home-vault-contents">
-      <div className="home-vault-contents-head">
-        <span className="home-vault-contents-label">In your vault</span>
-        <span className="home-vault-contents-path">{inv.root}</span>
-        <span className="home-vault-section-rule" />
-      </div>
-      <ul className="home-vault-contents-list">
-        {inv.entries.map((e) => (
-          <li key={e.name} className={`home-vault-row${!e.exists ? " home-vault-row--missing" : ""}`}>
-            <span className="home-vault-row-name">{e.label}</span>
-            <span className="home-vault-row-desc">
-              {e.description}
-              {e.meta && <span className="home-vault-row-meta"> · {e.meta}</span>}
-            </span>
-            <span className="home-vault-row-count">
-              {e.exists ? `${e.count.toLocaleString()} ${pluralize(e.count, e.unit)}` : "—"}
-            </span>
-            <span className="home-vault-row-size">
-              {/* Spaces dir on disk is always empty — real spaces have repo_path
-                  elsewhere — so its 0 B is meaningless. Show the vault total
-                  in that slot instead; it lands on the topmost row, reading
-                  as the headline number for the whole inventory. */}
-              {e.name === "spaces" ? formatBytes(inv.totalSize) : (e.exists ? formatBytes(e.size) : "")}
-            </span>
-          </li>
-        ))}
-      </ul>
-    </section>
-  );
-}
-
-function formatBytes(n: number): string {
-  if (n < 1024) return `${n} B`;
-  if (n < 1024 * 1024) return `${(n / 1024).toFixed(1)} KB`;
-  if (n < 1024 * 1024 * 1024) return `${(n / 1024 / 1024).toFixed(1)} MB`;
-  return `${(n / 1024 / 1024 / 1024).toFixed(2)} GB`;
-}
